front super-items: add the star ratings display
