+++ projects/Mallard/src/components/front/items/image-items.tsx
@@ -8,19 +8,14 @@ import {
     PropTypes,
     tappablePadding,
 } from './helpers/item-tappable'
-import {
-    getImageHeight,
-    isFullWidthItem,
-    isSmallItem,
-    isFullHeightItem,
-} from './helpers/sizes'
+import { isFullWidthItem, isSmallItem, isFullHeightItem } from './helpers/sizes'
 import { SportItemBackground } from './helpers/sports'
 import { TextBlock } from './helpers/text-block'
 import { SmallItem } from './small-items'
 import { Standfirst } from './helpers/standfirst'
 import { useIsOpinionCard, useIsSportCard } from './helpers/types'
-import { Stars } from 'src/components/stars/stars'
-import { TrailImage, ItemSizes } from 'src/common'
+import { TrailImageView } from './trail-image-view'
+import { getImageHeight } from './helpers/sizes'
 
 /*
 Normal img on top + text
@@ -52,17 +47,10 @@ const ImageItem = ({ article, size, ...tappableProps }: PropTypes) => {
     }
     return (
         <ItemTappable {...tappableProps} {...{ article }}>
-            {article.trailImage && (
-                <TrailImageView
-                    image={article.trailImage}
-                    itemSizes={size}
-                    starRating={
-                        article.type === 'article'
-                            ? article.starRating
-                            : undefined
-                    }
-                />
-            )}
+            <TrailImageView
+                article={article}
+                style={{ height: getImageHeight(size) }}
+            />
             {isSportCard && isFullWidthItem(size) ? (
                 <SportItemBackground
                     style={{
@@ -218,50 +206,6 @@ const SplitImageItem = ({ article, size, ...tappableProps }: PropTypes) => {
                 ) : null}
             </View>
         </ItemTappable>
-    )
-}
-
-const trailImageViewStyles = StyleSheet.create({
-    frame: {
-        width: '100%',
-        flex: 0,
-    },
-    image: {
-        width: '100%',
-        height: '100%',
-    },
-    rating: {
-        position: 'absolute',
-        backgroundColor: 'white',
-        left: 0,
-        bottom: 0,
-    },
-})
-
-/**
- * If there is a rating for the article (ex. a theater piece rating) then we
- * display it in the bottom-left corner of the image (by using absolute
- * positioning).
- */
-const TrailImageView = ({
-    image,
-    itemSizes,
-    starRating,
-}: {
-    image: TrailImage
-    itemSizes: ItemSizes
-    starRating?: number
-}) => {
-    const height = getImageHeight(itemSizes)
-    const frameStyle = [trailImageViewStyles.frame, { height }]
-    if (starRating == null) {
-        return <ImageResource style={frameStyle} image={image} />
-    }
-    return (
-        <View style={frameStyle}>
-            <ImageResource style={trailImageViewStyles.image} image={image} />
-            <Stars style={trailImageViewStyles.rating} rating={starRating} />
-        </View>
     )
 }
